\name{segBarData}
\alias{segBarData}
\docType{data}
\title{Test data to demonstrate segmented stacked bars}
\description{
   Test data used to demonstrate stacked bars: segmented, centered and normalized.
}
\usage{segBarData}
\format{
  This data sample contains five observations/measurements per state.  Each observation 
  will be represented by one segment of the stacked bar glyphic.\cr
  A data frame with one set of 5 observations per state::
  \describe{
    \item{X1}{a numeric vector of data.  The first observation of five.}
    \item{X2}{a numeric vector of data.  The second observation.}
    \item{X3}{a numeric vector of data.  The third observation.}
    \item{X4}{a numeric vector of data.  The fourth observation.}
    \item{X5}{a numeric vector of data.  The fourth observation.}
  }
  Each row has the state 2 character abbreviation as its row name.
}
\details{
   The dataset contains 51 records, one for each state containing 5 observations or 
   measurements.  In this example, the five measurements will be presented by the 
   stacked bar glyphic as a segmented, centered, or normalized stacked bar glyphic.
}
\keyword{datasets}
